+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 1.1
 
-- Expanded FAQ section
-- Added "Disable Camera Change While ___ is Held" setting (?)
-- Added "Disable Camera Change While in Combat" setting (?)
+- Added "Suppress Camera Change Key" setting
+- Added "Disable During Combat" setting
+- Added "Snap If Inventory Paused" setting
+- Removed "Restore Camera on Close" setting since it didn't wark well and didn't make much sense
+- Switched Select settings renderers with SuperSelect3 renderers by ownlyme. Now they are clickable
 - Fixed player rotating from the camera when using items from the inventory (?)
+- Fixed camera not changing when opening inventory in third person in combat/magic stance
+- Fixed all issues tied to OpenMW Camera builtin lua module
+- Simplified settings names and descriptions
 
 ## 1.0