+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 - The mod is now completely functional with paused inventory. Unpausers are no longer required, but still highly recommended
 - Updated settings
-- Switched Select settings renderers with SuperSelect3 renderers by ownlyme. Now they are clickable
-- Fixed player rotating from the camera when using items from the inventory (?)
+- Switched default Select settings renderers with SuperSelect3 by ownlyme. Now they are clickable
 - Fixed camera not changing when opening inventory in third person in combat/magic stance
 - Fixed all issues tied to OpenMW Camera builtin lua module